@@ -1,42 +1,35 @@
 {
-  "name": "musabhassan.com",
-  "author": {
-    "name": "Musab Hassan"
-  },
-  "version": "2.1.0",
-  "private": true,
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "rollup": "rollup"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.1.1",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "@tsconfig/svelte": "^3.0.0",
-    "rollup": "^2.62.0",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-glslify": "^1.2.1",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.49.0",
-    "svelte-check": "^2.2.11",
-    "svelte-preprocess": "^4.10.1",
-    "tslib": "^2.0.0",
-    "typescript": "^4.5.4"
-  },
-  "dependencies": {
-    "animejs": "^3.2.1",
-    "bezier-easing": "^2.1.0",
-    "detect-gpu": "^4.0.7",
-    "node-sass": "^8.0.0",
-    "sirv-cli": "^2.0.1",
-    "slickscrolljs": "^2.0.2",
-    "three": "^0.136.0",
-    "wait-for-element-transition": "^3.3.1"
-  }
+	"name": "musabhassan.com",
+	"author": {
+		"name": "Musab Hassan"
+	},
+	"version": "2.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+	},
+	"devDependencies": {
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@types/animejs": "^3.1.12",
+		"sass": "^1.69.7",
+		"svelte": "^4.2.7",
+		"svelte-check": "^3.6.0",
+		"tslib": "^2.4.1",
+		"typescript": "^5.0.0",
+		"vite": "^5.0.3"
+	},
+	"dependencies": {
+		"animejs": "^3.2.1",
+		"bezier-easing": "^2.1.0",
+		"detect-gpu": "^4.0.7",
+		"slickscrolljs": "^2.0.2",
+		"three": "^0.136.0"
+	},
+	"type": "module"
 }
