@@ -1,7 +1,7 @@
 import { imgPromises } from "./store";
 
 // Load images asynchronously during page loading animation
-export async function loadImage(src) {
+export async function loadImage(src: string) {
     const promise =  new Promise(async (resolve: (src: string) => void, reject) => {
         const blob = await (await fetch(src)).blob();
         const reader = new FileReader();
@@ -12,4 +12,12 @@ export async function loadImage(src) {
 
     imgPromises.update(val => [...val, promise]);
     return promise;
+}
+
+export function fetchJsonData(sourceFile: string) {
+    return new Promise(async (resolve: (data: any) => void) => {
+        const request = await fetch(sourceFile);
+        const data = await request.json();
+        resolve(data);
+    });
 }
